added new dao

+++ src/main/resources/db/migration/V1__create_table.sql
@@ -25,26 +25,26 @@ CREATE TABLE teachers(
     );
 
 CREATE TABLE classrooms(
-    classroom_id SERIAL NOT NULL PRIMARY KEY,
+    classroom_id BIGSERIAL NOT NULL PRIMARY KEY,
     classroom_number INTEGER NOT NULL
     );
 
 CREATE TABLE lesson_slot(
-    lesson_slot_id SERIAL NOT NULL PRIMARY KEY,
+    lesson_slot_id BIGSERIAL NOT NULL PRIMARY KEY,
     lesson_slot_number INTEGER NOT NULL
     );
 
 CREATE TABLE courses(
-    course_id SERIAL NOT NULL PRIMARY KEY,
+    course_id BIGSERIAL NOT NULL PRIMARY KEY,
     course_name VARCHAR(255) NOT NULL
     );
 
 CREATE TABLE classes(
     classes_id BIGSERIAL NOT NULL PRIMARY KEY,
-    "course_id" INTEGER REFERENCES courses (course_id),
+    "course_id" BIGINT REFERENCES courses (course_id),
     "teacher_id" BIGINT REFERENCES teachers (teacher_id),
     "group_id" BIGINT REFERENCES groups (group_id),
-    "classroom_id" INTEGER REFERENCES classrooms (classroom_id),
-    "lesson_slot_id" INTEGER REFERENCES lesson_slot (lesson_slot_id),
+    "classroom_id" BIGINT REFERENCES classrooms (classroom_id),
+    "lesson_slot_id" BIGINT REFERENCES lesson_slot (lesson_slot_id),
     classes_date BIGINT
 );
